Add PlayerMapper and convert Player to PlayerDto.
